+++ features/alerts/alertOnChumBucketAutopickup.js
@@ -32,7 +32,7 @@ export function playAlertOnBucketAutoPickup() {
 		Client.showTitle(`${YELLOW}Chum Bucket is gone`, '', 1, 30, 1);
 	
 		if (settings.soundMode !== OFF_SOUND_MODE) {
-            new Sound({ source: 'random.orb', volume: 1, pitch: 1 }).play();
+            World.playSound('random.orb', 1, 1);
         }
 	} catch (e) {
 		console.error(e);