+++ features/alerts/alertOnFishingFestivalEnded.js
@@ -92,7 +92,7 @@ function alertOnFestivalResults() {
         Client.showTitle(`${YELLOW}Fishing Festival ended`, '', 1, 30, 1);
 
         if (settings.soundMode !== OFF_SOUND_MODE) {
-            new Sound({ source: 'random.orb', volume: 1, pitch: 1 }).play();
+            World.playSound('random.orb', 1, 1);
         }
 
         const text = Object.entries(sharksCaught)