+++ features/alerts/alertOnNonFishingArmor.js
@@ -47,7 +47,7 @@ function alertOnNonFishingArmor(event) {
             Client.showTitle(`${RED}Equip fishing armor!`, '', 1, 25, 1);
     
             if (settings.soundMode !== OFF_SOUND_MODE) {
-                new Sound({ source: 'random.orb', volume: 1, pitch: 1 }).play();
+                World.playSound('random.orb', 1, 1);
             }
         }, 1000);    
     } catch (e) {