+++ features/alerts/alertOnThunderBottleCharged.js
@@ -24,7 +24,7 @@ function playAlertOnBottleCharged(bottleName) {
 		Client.showTitle(`${AQUA}${bottleName} is full`, '', 1, 30, 1);
 	
 		if (settings.soundMode !== OFF_SOUND_MODE) {
-            new Sound({ source: 'random.orb', volume: 1, pitch: 1 }).play();
+            World.playSound('random.orb', 1, 1);
         }
 	} catch (e) {
 		console.error(e);