+++ features/chat/messageOnHotspotFound.js
@@ -100,7 +100,7 @@ function announceFoundHotspot(position, perk) {
         ).chat();
     
         if (settings.soundMode !== OFF_SOUND_MODE) {
-            new Sound({ source: 'random.orb', volume: 1, pitch: 1 }).play();
+            World.playSound('random.orb', 1, 1);
         }
     }